refactor(ricer::cli): simplify module structure
